update all the events pages and create new ones
