+++ docs/events/background_events.md
@@ -3,7 +3,7 @@ import reflex as rx
 from pcweb import constants, styles
 from pcweb.base_state import State
 from pcweb.templates.docpage import docalert, doccode, docheader, subheader, docdemobox
-from pcweb.pages.docs.advanced_guide.background_tasks import LowLevelState, low_level_code, low_level_render_code, MyTaskState, my_task_code, my_task_render_code
+from pcweb.pages.docs.events.background_events import LowLevelState, low_level_code, low_level_render_code, MyTaskState, my_task_code, my_task_render_code
 ```
 
 # Background Tasks
