+++ pcweb/pages/docs/__init__.py
@@ -19,5 +19,6 @@ from .assets import *
 from .dynamic_rendering import *
 from .pages import *
 from .ui_overview import ui_overview
+from .events import *
 
 doc_routes = [r for r in locals().values() if isinstance(r, Route)]
