+++ pcweb/pages/docs/advanced_guide/__init__.py
@@ -1,4 +1,3 @@
-from .background_tasks import background_tasks
 from .custom_vars import custom_vars
 from .memoization import memoization
 from .use_middleware import use_middleware
