+++ pcweb/pages/docs/events/__init__.py
@@ -1,0 +1,8 @@
+from .background_events import background_events
+from .event_arguments import event_arguments
+from .events_overview import events_overview
+from .page_load_events import page_load_events
+from .setters import setters
+from .special_events_docs import special_events_docs
+from .yield_events import yield_events
+from .chaining_events import chaining_events
