+++ pcweb/pages/docs/events/event_arguments.py
@@ -1,0 +1,7 @@
+from pcweb import flexdown
+from pcweb.templates.docpage import docpage
+
+
+@docpage()
+def event_arguments():
+    return flexdown.render_file("docs/events/event_arguments.md")
