+++ pcweb/pages/docs/events/setters.py
@@ -1,0 +1,8 @@
+from pcweb import flexdown
+
+from pcweb.templates.docpage import docpage
+
+
+@docpage()
+def setters():
+    return flexdown.render_file("docs/events/setters.md")
