+++ pcweb/pages/docs/state/__init__.py
@@ -1,4 +1,3 @@
-from .events import events
 from .overview import state_overview
 from .substates import substates
 from .vars import vars
