+++ pcweb/pcweb.py
@@ -56,40 +56,40 @@ for route in doc_routes:
         image="/previews/docs_preview.png",
     )
 
+# # # Add the pages to the app.
+# for route in changelog_routes:
+#     app.add_page(
+#         route.component,
+#         route.path,
+#         route.title,
+#         description="Keep up to date with the latest Reflex news.",
+#         image="/previews/changelog_preview.png",
+#     )
+
 # # Add the pages to the app.
-for route in changelog_routes:
-    app.add_page(
-        route.component,
-        route.path,
-        route.title,
-        description="Keep up to date with the latest Reflex news.",
-        image="/previews/changelog_preview.png",
-    )
-
-# Add the pages to the app.
-for route in faq_routes:
-    app.add_page(
-        route.component,
-        route.path,
-        route.title,
-        description="Frequently asked questions about Reflex.",
-        image="/previews/faq_preview.png",
-    )
+# for route in faq_routes:
+#     app.add_page(
+#         route.component,
+#         route.path,
+#         route.title,
+#         description="Frequently asked questions about Reflex.",
+#         image="/previews/faq_preview.png",
+#     )
 
 
-for key in component_list:
-    for component_group in component_list[key]:
-        if isinstance(component_group[0], str):
-            continue
-        else:
-            path = f"/docs/library/{key.lower()}/{component_group[0].__name__.lower()}"
-            app.add_page(
-                multi_docs(path=path, component_list=component_group).component,
-                route=path,
-                title=component_group[0].__name__,
-                description=f"Reflex | Docs for {component_group[0].__name__} component.",
-                image="/previews/index_preview.png",
-            )
+# for key in component_list:
+#     for component_group in component_list[key]:
+#         if isinstance(component_group[0], str):
+#             continue
+#         else:
+#             path = f"/docs/library/{key.lower()}/{component_group[0].__name__.lower()}"
+#             app.add_page(
+#                 multi_docs(path=path, component_list=component_group).component,
+#                 route=path,
+#                 title=component_group[0].__name__,
+#                 description=f"Reflex | Docs for {component_group[0].__name__} component.",
+#                 image="/previews/index_preview.png",
+#             )
 
 # Add redirects
 redirects = [
